track VMC (vector magnitude count); predict periods of activity
and notify the Android app, which later can use that to try to
keep weather etc. up-to-date, in anticipation that the user might
need to know.  To avoid doing internet queries during periods of inactivity.

But also: ensure that the step count is accurate and persisted
at the transition to a new interval.