preparations for handling the rest of the forecast
Had an idea to merge precipitation with nowcast, but the trouble is that
the times are sent as an array of bytes, which means we can handle only times
+/- 128 minutes from now, or 256 if it's unsigned.  So not even a whole
day's worth, but it's just adequate for nowcast since that data only goes
2 hours in the future anyway.  And there's not a way to send an array of 16-bit
ints to the pebble, for example, AFAICT.  So I guess each forecast will have to
be its own dictionary with separate fields for time and whatever data.  So this
new mergeForecasts method probably won't be useful.

It's also a bit odd that Yr locationForecast XML has separate blocks for
regular forecast and precipitation, apparently because the latter is more
fine-grained.  So I should send it separately too because fine-grained
precipitation can be shown nicely on the watch face.  And I'd like to have
a second screen (accessible via a tap/shake gesture) which shows a linear
horizontal chart with precip. and temperature etc. like the yr.no site has.
That will go a bit further in the future.

Anyway I'm slowly modifying/getting rid of the imported "weathercast" code.

Pebble now avoids assuming that nowcast will always be less than 16 datapoints...
just in case.  (It was crashing when I first tried to merge precipitation,
because of that.)