use FetchService in the forecast widget too
It's not the same though... ForecastWidget should receive a broadcast
because there can be multiple instances.  But a broadcast intent cannot
be very large: the downloaded PNG data fits, a Bitmap does not.
So, each widget has to do its own PNG decoding for now.
This avoids direct access to the static bitmap variable in MeteogramService,
which is now obsolete.