handle weather independently from gisgraphy
Gisgraphy can be down, or slow, so don't wait for a response: fetch
the weather anyway.

Also try to send stored weather data to the Pebble rather than
waiting to fetch an update first.

Hopefully these two things will reduce the incidence of having nothing
to display on either end.

However, the Pebble still crashes a lot when using this with the official
Pebble app.  There are lots of nacks, too.