try harder to fix pebble comm/crash problems; other bug fixes
Don't assume that a succession of nacks means the watch is disconnected:
just stop overwhelming it with forecast data etc.  But still send the
current weather and location: those are the most important bits, and
also quite small.

Don't "resendEverything" on every tap and active interval.  On active
interval notification, ask for lazy weather update.  On tap, resent only
if nothing has been sent since the watch connected.

For some reason I never see the hello message now, when using the official
Pebble app, even though the watch tries to send it.

Don't send sunrise/sunset anymore: the watch is calculating it for itself,
which saves traffic and makes it more useful even when there is comm trouble.