Pebble power savings: don't send tap; repaint circle only every 5 minutes
No ticks per second, only per minute.
Even a disconnected Pebble Time Round otherwise is only getting a couple
of hours of battery life, it seems mainly because of trying to communicate
too much.  Turning Bluetooth off makes a big difference.  So we have to
reduce comms to the bare minimum, it seems.